quickpython/requirements.txt
pypdf==3.9.0
azure-ai-formrecognizer==3.2.1
azure-identity==1.13.0
feedparser==6.0.10
lxml==4.9.3
azure-ai-textanalytics==5.3.0
chardet==5.2.0
pathlib
argparse==1.4.0
uuid==1.30
azure.core
azure.cosmos
azure.keyvault
pandas
beautifulsoup4==4.10.0
azure-storage-blob == 12.18.2
azure-search-documents==11.4.0b8
rich == 12.5.1
openai==1.35.14